fix message length and crc8 calc

+++ OmniCore.Py/crc.cs
@@ -2,8 +2,8 @@ namespace OmniCore.py
 {
     public static class CrcUtil
     {
-        public static byte[] crc8_table =
-            new byte[]
+        public static ushort[] crc8_table =
+            new ushort[]
             {
                 0x00, 0x07, 0x0e, 0x09, 0x1c, 0x1b, 0x12, 0x15,
                 0x38, 0x3f, 0x36, 0x31, 0x24, 0x23, 0x2a, 0x2d,
